requirements.txt
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.10.0
defusedxml==0.7.1
dill==0.3.8
fonttools==4.51.0
idna==2.10
kiwisolver==1.4.5
Levenshtein==0.25.1
matplotlib==3.8.4
numpy==1.26.4
opencv-python-headless==4.8.0.74
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pydantic>=2,<3
PyYAML==6.0.1
rapidfuzz==3.9.0
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.28
scipy==1.13.0
six==1.16.0
supervision==0.20.0
tqdm==4.66.4
urllib3==2.2.1
seaborn
transformers
timm
einops
termcolor
markdown-it-py
